AboutService and ContactService: Add CRUD methods and ApiResponser trait
The code diff above shows the addition of CRUD methods and ApiResponser trait to the AboutService and ContactService classes. The new methods in each class include all, get, create, update, and delete methods for the corresponding models (About and Contact). The ApiResponser trait is used to return success or error responses with appropriate HTTP status codes.