"Refactored AboutController and ContactController to use AboutService and ContactService for business logic"
This commit refactors the AboutController and ContactController in the Api/About and Api/Contact directories respectively. The controllers have been updated to use the AboutService and ContactService for business logic related to getting and updating about and contact data. The constructor of each controller has been updated to inject the corresponding service. Additionally, methods for getting and updating about and contact data have been added to each controller, which delegate to the service methods.